refactor(runtime): lift skill-chain helpers into omicsclaw.runtime.skill_chain
bot/skill_orchestration.py owned three closely-related helpers
(_normalize_extra_args, _run_skill_via_shared_runner, _run_omics_skill_step)
that wrapped omicsclaw.core.skill_runner. None of them needed bot
state — only _run_omics_skill_step pulled OUTPUT_DIR from bot.core to
build a timestamped output directory. The trio was reverse-imported
from omicsclaw/runtime/preflight/sc_batch.py via lazy imports.

Move all three to omicsclaw/runtime/skill_chain.py with public names
(normalize_extra_args, run_skill_via_shared_runner, run_omics_skill_step).
run_omics_skill_step now takes an explicit output_root: Path so the
runtime side stays free of any bot-side path constant; the bot adapter
in bot/skill_orchestration.py keeps the OUTPUT_DIR default. The auto-
prepare chain in sc_batch.py accepts output_root and threads it to
the omicsclaw-side helper. bot/tool_executors.py passes OUTPUT_DIR at
the only call site.

Ratchet: GRANDFATHERED_LAZY_IMPORTS shrinks 14 → 13. The remaining
reverse import in sc_batch.py is the chain-tail execute_omicsclaw
call, handled in the next slice.
